Fix string corruption bug - Bump version to 0.06
Previously we tried to parse the string in place as we converted
it, however if the string is upgraded this can result in strangeness.
This patch fixes things so we are always reading the source string
and writing to the target string separately. If the target is upgraded
the read point is unaffected.